Move artifact manager
