+++ source/ArtifactManager.js
